Add files via upload
Linear Regression project: This project is scenario using real-world data to help you predict the percentages of Persons' who have high likelihood of potentially obtaining heart disease. 

Problem: 
as an expert data scientist, was given a historical data, in the “heart.csv” file, collected from
a health facility, which was monitoring the biking activities, smoking level, and the risk of having
a heart disease for many individuals. Since, sufficient data is available, the health care facility
decided to hire you to develop a machine learning model that can predict the risk of getting heart
disease by some people from biking activities and smoking level which are easily measurable. A
good predictive model may save millions of dollars to the facility.